+++ server/src/main/java/ee/zxz/helloapi/domain/DTO/ApiWeekArray.java
@@ -4,10 +4,10 @@ import java.time.LocalDateTime;
 
 public class ApiWeekArray {
     private LocalDateTime date;
-    private Integer count;
+    private Long count;
 
     public LocalDateTime getDate() { return date; }
     public void setDate(LocalDateTime date) { this.date = date; }
-    public Integer getCount() { return count; }
-    public void setCount(Integer count) { this.count = count; }
+    public Long getCount() { return count; }
+    public void setCount(Long count) { this.count = count; }
 }
